Major Improvements + Fixes
- Grimoire.csproj favicon finally added
- Grimoire.opt  data.clear_on_create setting added (For LamiaCore/InkDevil <3)
- GUI.DataRebuild.cs implemented
- GUI.DataRebuild.cs design finalized
- GUI.Main.cs Main_DragDrop event for Style.DATA updated to accomodate dropping files to be inserted
- GUI.MessageListBox.cs implemented for warning users about files to be inserted
- GUI.Settings.Designer.cs FormBorderStyle.cs set to SizeableToolWindow for proper user resizing
- Logs.Manager.cs Enter(Sender sender, Level level, string message, string stackTrace) overload removed for redundancy
- Properties.AssemblyInfo.cs file version updated
- Structures.Settings.cs ClearCreate setting implemented in settings GUI to take advantage of previously mentioned data.clear_on_create setting
- Tabs.Styles.Data.cs ts_file_new_Click event now enters log messages properly
- Tabs.Styles.Data.cs ts_file_new_Click event now disables core.Backups before attempting a new build (to avoid DataCore.core.createBackups errors about non-existant files)
- Tabs.Styles.Data.cs ts_file_new_Click event now properly clears DataCore.core or displays newly build client files after successful build
- Tabs.Styles.Data.cs ts_file_rebuild_Click event implemented to rebuild data.xxx files
- Tabs.Styles.Data.cs extensions_BeforeExpand event now properly sets extensions subnode text to avoid bugs seen in Gamerz bug report
- Tabs.Styles.Data.cs grid_SelectionChanged event now properly enables grid_cs.Items[2] (Insert) button
- Tabs.Styles.Data.cs grid_cs_insert_Click event upgraded to support multiple files, insert code moved to shared method private void insert_files(string[] paths)
- Tabs.Styles.Data.cs searchInput_TextChanged event removed || Searching which was causing searching the wrong dataset and hiding results (Bug report by Gamerz)
- Tabs.Styles.Data.cs unhook/hook_core_events methods implemented to accomodate implementation of GUI.DataRebuild.cs
- Tabs.Styles.Data.cs public Hook_Core_Events method implemented to allow GUI.RebuildData.cs to rehook events to the calling Data tab
- Tabs.Styles.Data.cs extension population and grid display trigger moved from finally {} clause to shared method display_data() so that ts_file_new_Click() event can trigger this code as-well
- Tabs.Styles.Data.cs display_data
 shared method implemented
- Tabs.Styles.Data.cs insert_files(string[] filePaths) implemented to accomodate file insert upgrades
- Tabs.Styles.Data.cs ts_file_new button now is hidden and ts_file_rebuild button exposed once a data.000 index has been loaded
- Tabs.Styles.RDB.cs now properly sets default data load directly when loading rdb via 'Data' option
- Utilities.Paths.cs string[] filePaths implemented to allow multiple file selection without FolderBrowserDialog
- Utilities.Paths.cs FilePath/FilePaths implemented to accomodate upgrades to the File Import features of Tabs.Styles.Data.cs
- Utilities.StringEXT.cs updated to allow showing decimal style variables instead of int only
- README.md updated to reflect current status

+++ Utilities/StringExt.cs
@@ -8,7 +8,7 @@ namespace Grimoire.Utilities
 
         public static string FormatToSize(long length)
         {
-            long len = length;
+            double len = length;
             int order = 0;
 
             while (len >= 1024 && order < sizes.Length - 1)
